🔧 Make minor improvements

+++ os/install_npm_packages.sh
@@ -10,6 +10,12 @@ declare -r -a NPM_PACKAGES=(
 
 main() {
 
+    # Check if `npm` is installed
+    if [ $(cmd_exists "npm") -eq 1 ]; then
+        print_error "npm is required, please install it!\n"
+        exit
+    fi
+
     # Make sure the most recent version of `npm` is installed
     execute "npm install --global npm" "npm (update)"
